soap-service/requirements.txt
Flask
psycopg2-binary
zeep

